Home page and forms done desing

+++ frontend/app/components/FormSignUp.tsx
@@ -8,16 +8,25 @@ export default function Form({ route }: { route: string }) {
   const [password, setPassword] = useState("");
   const [username, setUsername] = useState("");
   const [loading, setLoading] = useState(false);
+  const [passwordconfirm, setPasswordconfirm] = useState("");
+  const [error, setError] = useState("");
   const router = useRouter();
 
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     setLoading(true);
     e.preventDefault();
-    router.push("/login");
+    if (password != passwordconfirm) {
+      setError("Passwords do not match");
+      setLoading(false);
+      return;
+    }
     try {
       const res = await api.post(route, { username, email, password });
+      router.push("/login");
     } catch (error) {
-      alert(error);
+      setError(
+        "There is an existing user with this email or username. Please try again."
+      );
     } finally {
       localStorage.clear();
       setLoading(false);
@@ -27,32 +36,55 @@ export default function Form({ route }: { route: string }) {
   return loading ? (
     <Loading />
   ) : (
-    <form onSubmit={handleSubmit} className="form-container">
-      <h1>Sign Up</h1>
-      <input
-        className="form-input"
-        type="text"
-        value={username}
-        onChange={(e) => setUsername(e.target.value)}
-        placeholder="Username"
-      ></input>
-      <input
-        className="form-input"
-        type="email"
-        value={email}
-        onChange={(e) => setEmail(e.target.value)}
-        placeholder="Email"
-      ></input>
-      <input
-        className="form-input"
-        type="password"
-        value={password}
-        onChange={(e) => setPassword(e.target.value)}
-        placeholder="Password"
-      ></input>
-      <button className="form-button" type="submit">
-        Sign Up
-      </button>
-    </form>
+    <div className="form-wrapper">
+      <form onSubmit={handleSubmit} className="form-container">
+        {error && <div className="alert alert-danger">{error}</div>}
+        <h1 className="label1">Sign Up</h1>
+        <div className="mb-3">
+          <label className="form-label">Username</label>
+          <input
+            type="username"
+            className="form-control"
+            value={username}
+            onChange={(e) => setUsername(e.target.value)}
+            placeholder="Enter your username"
+          />
+        </div>
+        <div className="mb-3">
+          <label className="form-label">Email address</label>
+          <input
+            type="email"
+            className="form-control"
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+            placeholder="Enter email"
+          />
+        </div>
+        <div className="mb-3">
+          <label className="form-label">Password</label>
+          <input
+            type="password"
+            className="form-control"
+            value={passwordconfirm}
+            onChange={(e) => setPasswordconfirm(e.target.value)}
+            placeholder="Confirm Password"
+          />
+        </div>
+        <div className="mb-3">
+          <label className="form-label">Confirm Password</label>
+          <input
+            type="password"
+            className="form-control"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            placeholder="Password"
+          />
+        </div>
+        <a href="/login">Already have an account?Login</a>
+        <button type="submit" className="btn btn-dark w-100">
+          Sign Up
+        </button>
+      </form>
+    </div>
   );
 }
